Debugging for a probable AVX2-related rustc bug
This commit contains `dbg!` statements which are narrowed down to be as
close to the problem as possible.

The following commands succeed:

- `cargo test donna_self_test1 -- --test-threads=1 --nocapture`
- `RUSTFLAGS="-Ctarget-cpu=haswell" cargo test donna_self_test1 --release -- --test-threads=1 --nocapture`

The following command FAILS with a miscomputed output:

- `cargo test donna_self_test1 --release -- --test-threads=1 --nocapture`

Included in this commit are `dbg!` statements capturing the miscompiled
output and comments with instructions about small tweaks which make the
tests succeed under the conditions where it presently fails.

These include:

- Inserting a `dbg!` statement in the right place
- Commenting out a `#[target_feature(enable = "avx2")]` attribute